Change stardate affinity type to FLOAT, so that “.0” stardates get stored and displayed exactly as such.
FossilOrigin-Name: fb1c9eba382e4424b8ec782cc597d7cde72201fb73592e87138a4b57240eed41

+++ schema.sql
@@ -39,7 +39,7 @@ CREATE TABLE episode (
        season INTEGER,
        episode_number INTEGER,
        production_code TEXT,
-       stardate NUMERIC,
+       stardate FLOAT,
        date DATETIME,
        vignette BOOLEAN
 );
